fix(amplify): Prevent deployment failures when setting stage field
Problem:
The `stage` field on both the `App` and `Branch` constructs is modeled
as an open ended string. However, the control plane has validation that
enforces using one of the following values:
* DEVELOPMENT
* BETA
* EXPERIMENTAL
* PULL_REQUEST
* PRODUCTION
When not set to one of these values, the CFN deployment will fail.

Solution:
Creating an enum to represent the valid values was considered, however,
the only value that changes the system's behavior is PRODUCTION which
adds the `productionBranch` object on the response retuned from
`GetApp`. This object is used to label a branch as the "production"
branch in the AWS Amplify Console, but not used anywhere else. The
service team considers the `stage` field all but deprecated and would
like to move away from App level Branch configurations.

Therefore, the best experience is to provide a flag that limits the
possible values passed to `stage` to just `PRODUCTION` or undefined.

Testing Done:
* `yarn build+test+extract`